Created a battle system
During debugging, when creating the player, the program seems to have an issue with the move EatUp. Following the order of the code, I have discovered that the Item Apple gets past its super constructor, but then crashes. I don't know if the crash happens because of the next loaded Item, that being fish, or if it happens right after the successful loading of apple. Regardless, fish doesn't get past its super constructor